Local optimization

+++ resources/test-template.py
@@ -3,9 +3,12 @@ import shutil
 import subprocess
 import time
 
-for cmd in ["docker container prune -f",
-            "docker image rm example-site:latest example-nginx:latest -f"]:
-  subprocess.check_call(cmd, shell=True)
+onGithubActionMachine = os.path.exists(os.path.join("/", "home", "runner"))
+
+if not onGithubActionMachine:
+    for cmd in ["docker container prune -f",
+                "docker image rm example-site:latest example-nginx:latest -f"]:
+      subprocess.check_call(cmd, shell=True)
 
 # ------------------------------
 print("Updating environment...")
@@ -27,11 +30,15 @@ if (not os.path.exists(environmentFile)):
 print("Spawning containers...")
 # -----------------------------
 subprocess.check_call("docker-compose up -d", shell=True)
-
 # TODO: When the container gets a new enough version of docker, can just use
 # `docker compose up -d --wait` [although not quite now that we need to wait
 # for the database to spin up]
-time.sleep(30)
+
+# For now, just sleep until everything is probably done with initial setup
+sleepTime = 30 if onGithubActionMachine else 10
+if (os.path.exists(os.path.join("home", "runner"))):
+    time.sleep(30)
+time.sleep(10)
 
 subprocess.check_call("docker-compose logs", shell=True)
 
@@ -44,8 +51,10 @@ subprocess.check_call("python3 runtests.py", shell=True)
 print("Cleaning up...")
 # ---------------------
 subprocess.check_call("docker-compose down", shell=True)
-os.remove(environmentFile)
-for cmd in ["docker container prune -f",
-            "docker image rm example-site:latest example-nginx:latest -f"]:
-  subprocess.check_call(cmd, shell=True)
+
+if not onGithubActionMachine:
+    os.remove(environmentFile)
+    for cmd in ["docker container prune -f",
+                "docker image rm example-site:latest example-nginx:latest -f"]:
+        subprocess.check_call(cmd, shell=True)
